requirements.txt
appdirs==1.4.3
Django==1.11.2
django-cors-headers==2.1.0
djangorestframework==3.6.3
djangorestframework-gis==0.11.2
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
packaging==16.8
psycopg2==2.7.1
PyJWT==1.5.3
pyparsing==2.2.0
pytz==2017.2
six==1.10.0